requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
motor==3.1.2
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
loguru==0.7.0
fastapi-pagination==0.12.4
email-validator==2.0.0 